Working Copy Stage One
Dresses can be added and stored. Each user has a profile view page, and a list of their associated dresses. The login window is now modal. Add Dress is also a modal view and the button for adding a dress is appended to the bottom of the dress list rather than statically on the navigation bar (which no longer exists). The navigation bar was redundant as page descriptions are given in the tab bar view controller, and the add dress button is not used very frequently. Also, its removal provides more space, especially on 3.5” form factor screens.